requirements.txt
psycopg2-binary>=2.9,<3.0
geoip2>=4.0,<5.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
